Prepare file to standardise Ui class
