+++ src/main/java/seedu/duke/command/ProjectCommand.java
@@ -2,7 +2,7 @@ package seedu.duke.command;
 
 import seedu.duke.exception.DukeException;
 import seedu.duke.project.Project;
-import seedu.duke.ui.Ui;
+import seedu.duke.ui.old.Ui;
 
 import java.util.ArrayList;
 import java.util.Hashtable;
