+++ src/main/java/seedu/duke/command/sprint/SprintCommand.java
@@ -2,7 +2,7 @@ package seedu.duke.command.sprint;
 
 import seedu.duke.command.Command;
 import seedu.duke.model.Project;
-import seedu.duke.ui.TextUi;
+import seedu.duke.ui.Ui;
 
 import java.util.Hashtable;
 
@@ -21,6 +21,6 @@ public abstract class SprintCommand extends Command {
      * @param ui UI that handles user interaction
      * @return Boolean - True if Bye command is executed
      */
-    public abstract boolean execute(Project proj, TextUi ui);
+    public abstract boolean execute(Project proj, Ui ui);
 
 }
