+++ src/main/java/seedu/duke/model/ProjectMembers.java
@@ -1,14 +1,12 @@
 package seedu.duke.model;
 
-import seedu.duke.Member;
-
 import java.util.ArrayList;
 import java.util.List;
 
-public class ProjectMember {
+public class ProjectMembers {
     public ArrayList<Member> memberList;
 
-    public ProjectMember() {
+    public ProjectMembers() {
         memberList = new ArrayList<>(100);
     }
